update Moved objects to single file. Update
