+++ mule/src/offline/training00_utilities.py
@@ -43,12 +43,65 @@ import tqdm
 #from IPython import get_ipython
 #import cv2
 
+
+import glob
+#import json
+#import pandas as pd
+#import tensorflow as tf
+# Check versions
+#assert tf.__version__ == '1.8.0'
+
+import logging
+#import zipfile
+#import re
+#import datetime
+#import cv2
+#import shutil
+#import json
+#from tabulate import tabulate
+#import tqdm
+#from IPython import get_ipython
+import cv2
+
+#%%
+class NoPlots:
+    def __enter__(self):
+        get_ipython().run_line_magic('matplotlib', 'qt')
+        plt.ioff()
+    def __exit__(self, type, value, traceback):
+        get_ipython().run_line_magic('matplotlib', 'inline')
+        plt.ion()
+
+
 #%% Logging
 #>>> import warnings
 #>>> image = np.array([0, 0.5, 1], dtype=float)
 #>>> with warnings.catch_warnings():
 #...     warnings.simplefilter("ignore")
 #...     img_as_ubyte(image)
+
+
+
+
+#%% LOGGING for Spyder! Disable for production. 
+logger = logging.getLogger()
+logger.handlers = []
+
+# Set level
+logger.setLevel(logging.DEBUG)
+
+# Create formatter
+#FORMAT = "%(asctime)s - %(levelno)s - %(module)-15s - %(funcName)-15s - %(message)s"
+#FORMAT = "%(asctime)s L%(levelno)s: %(message)s"
+FORMAT = "%(asctime)s - %(levelname)s - %(funcName) -20s: %(message)s"
+DATE_FMT = "%Y-%m-%d %H:%M:%S"
+formatter = logging.Formatter(FORMAT, DATE_FMT)
+
+# Create handler and assign
+handler = logging.StreamHandler(sys.stderr)
+handler.setFormatter(formatter)
+logger.handlers = [handler]
+logger.critical("Logging started")
 
 class LoggerCritical:
     def __enter__(self):
@@ -58,15 +111,42 @@ class LoggerCritical:
         my_logger = logging.getLogger()
         my_logger.setLevel("DEBUG")
 
+#
+#
+#
+#
+#
+#class LoggerCritical:
+#    def __enter__(self):
+#        my_logger = logging.getLogger()
+#        my_logger.setLevel("CRITICAL")
+#    def __exit__(self, type, value, traceback):
+#        my_logger = logging.getLogger()
+#        my_logger.setLevel("DEBUG")
+#
+#
+#import logging
+#logger = logging.getLogger()
+#logger.setLevel(logging.DEBUG)
+#logging.debug("test")
+#
+#with LoggerCritical():
+#    logging.debug("test block")
+#%%
+def remove_outliers(this_series):
+    """Given a pd.Series, return a new Series with no outliers
+    """
+    no_outlier_mask = np.abs(this_series-this_series.mean()) <= (3*this_series.std())
+    return this_series[no_outlier_mask]
 
-import logging
-logger = logging.getLogger()
-logger.setLevel(logging.DEBUG)
-logging.debug("test")
 
-with LoggerCritical():
-    logging.debug("test block")
+#%%
 
+def mm2inch(value):
+    return value/25.4
+PAPER_A3_LAND = (mm2inch(420),mm2inch(297))
+PAPER_A4_LAND = (mm2inch(297),mm2inch(210))
+PAPER_A5_LAND = (mm2inch(210),mm2inch(148))
 #%%
     
 
@@ -86,99 +166,98 @@ def splitall(path):
     return allparts    
 
 
-#%% DATAGEN
-class MuleDataGenerator(ks.utils.Sequence):
-    """Generates data for Keras"""
-    def __init__(self, indices, dataset, 
-                 batch_size=32, dim=None, n_channels=None, n_classes=15, shuffle=True):
-        """Keras data generator
-        
-        Aggregates the AIDataSet class
-        
-        Attributes:
-            indices (str): The allowed timestamps for data generation
-            dataset (AIDataSet): The dataset object with it's df and npz
-            batch_size : 
-            dim : 
-            n_channels : 
-            n_classes :
-            shuffle :
-        """
-        self.indices = indices
-        self.dataset = dataset
-        self.batch_size = batch_size
-        self.dim = dim
-        self.n_channels = n_channels
-        self.n_classes = n_classes
-        self.shuffle = shuffle
-        self.on_epoch_end()
-        
-        logging.debug("** Initialize datagen **".format())
-        logging.debug("Data folder: {}".format(dataset.data_folder))
-        
-        logging.debug("{} of {} total records used for generation".format(len(self.indices), len(self.dataset.df)))
-        #logging.debug("Frames NPZ located at: {}".format(self.dataset.path_frames_npz))
-        logging.debug("{} samples over batch size {} yields {} batches".format(len(self.indices),
-                                                                                   self.batch_size,
-                                                                                   math.ceil(len(self.indices)/self.batch_size),))
-        
-    def __len__(self):
-        """Keras generator method - Denotes the number of batches per epoch
-        """        
-        return int(np.floor(len(self.indices) / self.batch_size))
+#%%
+
+class VideoWriter:
+    """From a folder containing jpg images - create a video. 
     
-    # GET A BATCH!
-    def __getitem__(self, index): 
-        """Keras generator method - Generate one batch of data
-        """         
-        logging.debug("Generating batch {}".format(index))
+    
+    """
+    
+    def __init__(self,jpg_folder, path_vid_out, fps):
+        self.jpg_folder = jpg_folder
+        self.path_vid_out = path_vid_out
+        self.fps = fps
+        assert os.path.exists(self.jpg_folder), "{} does not exist".format(self.jpg_folder)
         
-        # Generate indexes of the batch
-        batch_indices = self.indices[index*self.batch_size:(index+1)*self.batch_size]
+        jpg_files = glob.glob(os.path.join(self.jpg_folder,'*.jpg'))
+        
+        logging.debug("{} video JPG frames found in {}".format(len(jpg_files),self.jpg_folder))
 
-        # Generate data by selecting these IDs
-        X, y = self.__data_generation(batch_indices)
+        self.jpg_files = self.sort_jpgs(jpg_files)
+        
+        logging.debug("Output video set to {} at {} fps".format(self.path_vid_out,self.fps))
 
-        return X, y
+        self.height, self.width = self.get_dimensions(self.jpg_files[0])
 
-    def on_epoch_end(self):
-        """Keras generator method - Shuffles indices after each epoch
+    def get_dimensions(self,this_jpg_path):
+        # Load a single frame to get dimensions
+        img_arr = mpl.image.imread(this_jpg_path)
+        frames_height = img_arr.shape[0]
+        frames_width = img_arr.shape[1]
+        logging.debug("Dimensions: {} x {} pixels (Height x Width)".format(frames_height,frames_width))
+        
+        return frames_height,frames_width
+
+    def sort_jpgs(self,jpg_files):
+        """Sort on file name (timestamp)
+        
         """
-        #self.indexes = np.arange(len(self.indices))
-        if self.shuffle == True:
-            # Shuffle is in-place! 
-            np.random.shuffle(self.indices)
+        frame_paths = list()
+        for this_img_path in jpg_files:
+            this_frame_dict = dict()
+            _, this_img_fname = os.path.split(this_img_path)
+            timestamp = os.path.splitext(this_img_fname)[0]
+            this_frame_dict['timestamp'] = timestamp
+            this_frame_dict['path'] = this_img_path
+            frame_paths.append(this_frame_dict)
+
+        # Sort the frames!
+        frame_paths_sorted = sorted(frame_paths, key=lambda k: k['timestamp'])
+        logging.debug("Sorted {} video frame image paths".format(len(frame_paths_sorted)))
+        
+        return [fd['path'] for fd in frame_paths_sorted]
+    
+    def write_video(self, num_frames=None, overwrite=False):
+        """From a list of frame JPG paths, generate a MP4. 
+        
+        Optionally specify the length of the video in num_frames (good for testing)
+        """
+        
+        if os.path.exists(self.path_vid_out) and not overwrite:
+            logging.debug("{} video already exists, skip unless overwrite=True".format(self.path_vid_out))
             
-    def __get_npy_arrays(self,batch_indices):
-        """Custom method - get the X input arrays
+            return
         
-        Open the npz file and load n frames into memory
-        """
-        # This is a pointer to the file
-        npz_file=np.load(self.dataset.path_frames_npz)
+        # This is extremely picky, and can fail (create empty file) with no warning !!
+        writer = cv2.VideoWriter(self.path_vid_out, cv2.VideoWriter_fourcc(*"MJPG"), self.fps, (self.width,self.height))
+        if not num_frames:
+            frames_to_write = self.jpg_files
+        else:
+            frames_to_write = self.jpg_files[0:num_frames]
+        with NoPlots(), LoggerCritical():
+            for this_jpg_path in tqdm.tqdm(frames_to_write):
+                img_arr = mpl.image.imread(this_jpg_path)
+                writer.write(img_arr) # Write out frame to video
         
-        frames_array = np.stack([npz_file[idx] for idx in batch_indices], axis=0)
-        logging.debug("Generating {} frames: {}".format(frames_array.shape[0], frames_array.shape))
+        logging.debug("Wrote {} frames to {}".format(len(frames_to_write),self.path_vid_out))
         
-        return frames_array
+        writer.release()
+        cv2.destroyAllWindows()
     
-    def __get_records(self,batch_indices):
-        """Custom method - get the y labels
-        """
-        this_batch_df = self.dataset.df.loc[batch_indices]
-        steering_values = this_batch_df['steering_signal'].values
-        steering_records_array = self.dataset.bin_Y(steering_values)
-        logging.debug("Generating {} records {}:".format(steering_records_array.shape[0],steering_records_array.shape))
-        return steering_records_array
+    def test_write(self):
+        """To test cv2 import, dimensions (cv2 is very picky), etc.
         
-    def __data_generation(self, batch_indices):
-        """Keras generator method - Generates data containing batch_size samples
+        Write some random noise to the video out path. 
         """
-
-        X = self.__get_npy_arrays(batch_indices)
-        y = self.__get_records(batch_indices)
-
-        return X, y
-  
-
+        
+        # Dimensions, for testing purposes
+        H = 480
+        W = 640
+        writer = cv2.VideoWriter(self.path_vid_out, cv2.VideoWriter_fourcc(*"MJPG"), 30, (W,H))
+        for frame in tqdm.tqdm(range(400)):
+            this_frame = np.random.randint(0, 255, (H,W,3)).astype('uint8')
+            writer.write(this_frame)
+        writer.release()        
+        logging.debug("Wrote test video to {}".format(self.path_vid_out))
